+++ src/test/java/com/kt/service/AccountServiceTest.java
@@ -300,109 +300,4 @@ class AccountServiceTest {
 
 	}
 
-	@Test
-	void 관리자_다른_계정_비밀번호_초기화_성공() {
-		PasswordRequestEntity passwordRequest = PasswordRequestEntity.create(
-			member1,
-			null,
-			PasswordRequestType.RESET
-		);
-		passwordRequestRepository.save(passwordRequest);
-		String originPassword = "1234";
-
-		assertNotNull(passwordRequest.getId());
-		accountService.resetAccountPassword(passwordRequest.getId());
-
-		log.info(
-			"isMatch :: {}", passwordEncoder.matches(
-				originPassword, member1.getPassword()
-			)
-		);
-
-		assertFalse(
-			passwordEncoder.matches(
-				originPassword,
-				member1.getPassword()
-			)
-		);
-		log.info("passwordRequest status : {}", passwordRequest.getStatus());
-	}
-
-	@Test
-	void 관리자_다른_계정_비밀번호_초기화_실패_요청사항_없음() {
-
-		assertThatThrownBy(
-			() -> accountService.resetAccountPassword(member1.getId())
-		).isInstanceOf(CustomException.class);
-
-	}
-
-
-	@Test
-	void 관리자_다른_계정_비밀번호_변경_성공() {
-		String originPassword = "1234";
-		String updatedPassword = "1231231!";
-		PasswordRequestEntity passwordRequest = PasswordRequestEntity.create(
-			member1,
-			updatedPassword,
-			PasswordRequestType.UPDATE
-		);
-		passwordRequestRepository.save(passwordRequest);
-		assertNotNull(passwordRequest.getId());
-		accountService.updateAccountPassword(passwordRequest.getId());
-
-		assertFalse(
-			passwordEncoder.matches(originPassword, member1.getPassword())
-		);
-
-		log.info(
-			"isMatch :: {}", passwordEncoder.matches(
-				originPassword, member1.getPassword()
-			)
-		);
-		assertEquals(PasswordRequestStatus.COMPLETED, passwordRequest.getStatus());
-		log.info("passwordRequest status : {}", passwordRequest.getStatus());
-	}
-
-	@Test
-	void 비밀번호_변경_및_초기화_요청_리스트_조회_성공() {
-		String updatePassword = "123123";
-
-		PasswordRequestEntity firstRequest = PasswordRequestEntity.create(
-			member1,
-			null,
-			PasswordRequestType.RESET
-		);
-
-		PasswordRequestEntity secondRequest = PasswordRequestEntity.create(
-			courier1,
-			updatePassword,
-			PasswordRequestType.UPDATE
-		);
-		passwordRequestRepository.save(firstRequest);
-		passwordRequestRepository.save(secondRequest);
-
-		Pageable pageable = Pageable.ofSize(10);
-
-		// when
-		PasswordRequest.Search request = new PasswordRequest.Search(
-			UserRole.COURIER,
-			null,
-			null,
-			""
-		);
-
-		Page<PasswordRequestResponse.Search> result =
-			accountService.searchPasswordRequests(request, pageable);
-
-		// then
-		assertThat(result).isNotNull();
-		assertThat(result.getContent()).hasSize(1);
-		assertThat(result.getContent()
-			.stream()
-			.map(PasswordRequestResponse.Search::accountId)
-		).contains(courier1.getId());
-
-	}
-
 }
